+++ flink-cdc-cli/src/test/resources/definitions/pipeline-definition-full.yaml
@@ -47,7 +47,7 @@ transform:
     partition-keys: product_name
     table-options: comment=app order
     description: project fields from source table
-    post-transform-converter: SOFT_DELETE
+    converter-after-transform: SOFT_DELETE
   - source-table: mydb.web_order_.*
     projection: CONCAT(id, order_id) as uniq_id, *
     filter: uniq_id > 10
