+++ flink-cdc-runtime/src/main/java/org/apache/flink/cdc/runtime/operators/transform/converter/PostTransformConverter.java
@@ -15,26 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.flink.cdc.common.transform.converter;
+package org.apache.flink.cdc.runtime.operators.transform.converter;
 
+import org.apache.flink.cdc.common.annotation.Experimental;
 import org.apache.flink.cdc.common.event.DataChangeEvent;
 
+import java.io.Serializable;
 import java.util.Optional;
 
-import static org.apache.flink.cdc.common.event.OperationType.DELETE;
+/**
+ * The PostTransformConverter applies to convert the {@link DataChangeEvent} after other part of
+ * TransformRule in PostTransformOperator.
+ */
+@Experimental
+public interface PostTransformConverter extends Serializable {
 
-/** This {@link PostTransformConverter} convert delete events to insert events. */
-public class SoftDeleteConverter implements PostTransformConverter {
-
-    @Override
-    public Optional<DataChangeEvent> convert(DataChangeEvent dataChangeEvent) {
-        if (DELETE.equals(dataChangeEvent.op())) {
-            return Optional.of(
-                    DataChangeEvent.insertEvent(
-                            dataChangeEvent.tableId(),
-                            dataChangeEvent.before(),
-                            dataChangeEvent.meta()));
-        }
-        return Optional.of(dataChangeEvent);
-    }
+    /** Change some parts of {@link DataChangeEvent}, like op or meta. */
+    Optional<DataChangeEvent> convert(DataChangeEvent dataChangeEvent);
 }
