+++ flink-cdc-runtime/src/main/java/org/apache/flink/cdc/runtime/operators/transform/converter/PostTransformConverters.java
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.flink.cdc.common.transform.converter;
+package org.apache.flink.cdc.runtime.operators.transform.converter;
 
 import org.apache.flink.cdc.common.annotation.Experimental;
+import org.apache.flink.cdc.common.utils.StringUtils;
 
 import java.util.Optional;
 
@@ -26,13 +27,18 @@ import java.util.Optional;
 public class PostTransformConverters {
     public static final String SOFT_DELETE_CONVERTER = "SOFT_DELETE";
 
-    public static Optional<PostTransformConverter> getInternalPostTransformConverter(
-            String identifier) {
+    /** Get the {@link PostTransformConverter} by given identifier. */
+    public static Optional<PostTransformConverter> of(String identifier) {
+        if (StringUtils.isNullOrWhitespaceOnly(identifier)) {
+            return Optional.empty();
+        }
+
         switch (identifier) {
             case SOFT_DELETE_CONVERTER:
                 return Optional.of(new SoftDeleteConverter());
             default:
-                return Optional.empty();
+                throw new IllegalArgumentException(
+                        String.format("Failed to find the converter %s.", identifier));
         }
     }
 }
